+++ CHANGELOG.md
@@ -65,7 +65,20 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- The `ex.Engine` had a new `enableCanvasContextMenu` arg that can be used to enable the right click context menu, by default the context menu is disabled which is what most games seem to want.
+- Added a method to force graphics on screen `ex.GraphicsComponent.forceOnScreen`
+- Added new `ex.Slide` scene transition, which can slide a screen shot of the current screen: `up`, `down`, `left`, or `right`. Optionally you can add an `ex.EasingFunction`, by default `ex.EasingFunctions.Linear`
+  ```typescript
+  game.goToScene('otherScene', {
+    destinationIn: new ex.Slide({
+      duration: 1000,
+      easingFunction: ex.EasingFunctions.EaseInOutCubic,
+      slideDirection: 'up'
+    })
+  });
+  ```
+- Added inline SVG image support `ex.ImageSource.fromSvgString('<svg>...</svg>')`, note images produced this way still must be loaded.
+- Added ability to optionally specify sprite options in the `.toSprite(options:? SpriteOptions)`
+- The `ex.Engine` constructor had a new `enableCanvasContextMenu` arg that can be used to enable the right click context menu, by default the context menu is disabled which is what most games seem to want.
 - Child `ex.Actor` inherits opacity of parents
 - `ex.Engine.timeScale` values of 0 are now supported
 - `ex.Trigger` now supports all valid actor constructor parameters from `ex.ActorArgs` in addition to `ex.TriggerOptions`
@@ -105,6 +118,9 @@ are doing mtv adjustments during precollision.
 
 ### Fixed
 
+- Fixed issue where `ex.Engine.screenshot()` images may not yet be loaded in time for use in `ex.Transition`s
+- Fixed issue where there would be an incorrect background color for 1 frame when transitioning to a new scene
+- Fixed issue where `blockInput: true` on scene transition only blocked input events, not accessors like `wasHeld(...)` etc.
 - Fixed issue where users could not easily define a custom `RendererPlugin` because the type was not exposed
 - Fixed issue where `ex.Fade` sometimes would not complete depending on the elapsed time
 - Fixed issue where `ex.PolygonColliders` would get trapped in infinite loop for degenerate polygons (< 3 vertices)
@@ -160,6 +176,7 @@ are doing mtv adjustments during precollision.
 ### Changed
 
 - Applied increased TS strictness:
+  * Director API subtree
   * Resource API subtree
   * Graphics API subtree
   * TileMap API subtree